[Fix] #38 - TestRoomAddressEnterView 코드정리 및 SwiftData관련 코드 삭제
